Clean various tiny things up

+++ src/collision.cpp
@@ -136,11 +136,11 @@ int x_intercept(line_seg_t *line, int y_pos) {
 void process_collisions() {
     for(auto *it = PhysicsBody::objects.begin(); it < PhysicsBody::objects.end();) {
         PhysicsBody *old_ptr = *it;
-        uint bottom_y = (**it).position_y + (**it).height;
+        int bottom_y = (**it).position_y + (int)(**it).height;
         for(auto *other = it + 1; other < PhysicsBody::objects.end() && (**other).position_y <= bottom_y;) {
             PhysicsBody *old_other_ptr = *other;
-            if((**other).position_x < (**it).position_x + (**it).width &&
-               (**it).position_x < (**other).position_x + (**other).width) {
+            if((**other).position_x < (**it).position_x + (int)(**it).width &&
+               (**it).position_x < (**other).position_x + (int)(**other).width) {
                 (**it).handle_collision(*other);
                 if(old_ptr != *it) break;
             }
